Small cleanup after introducing inheritance into temporal types
